s390x: Re-implement STFLE as an extension
The existing implementation of the STFLE instruction does not use the
correct operand size when tracking memory effects.  Instead of respecting
the user-provided maximum number of doublewords and the returned value
from the instruction, it assumes a hard coded value (S390_NUM_FACILITY_DW)
instead.

For example, if an application passes a buffer of 3 doublewords to STFLE
while Valgrind assumes a fixed size of 4 doublewords, Valgrind may falsely
complain about an invalid write for the last doubleword.

Fix this by re-implementing STFLE via the extension mechanism.