s390x: Fix disassembly of locfh/locfhr, update S390_MAX_MNEMONIC_LEN
The length of the "longest mnemonic" for the s390x disassembler is
currently defined in s390_defs.h to be 8 characters, where in fact it
should be 9.  Update the constant to reflect that.

Also fix the disassembly of the instructions locfh and locfhr, changing
them from their current wrong representations `locgh' and `locghr'.