Bug 493970 - s390x: Drop saving/restoring FPC upon helper call
Saving the FPC before each helper call and restoring it afterwards creates
unnecessary overhead, and it may also not be desirable.

Drop it.  Also remove the functions in host_s390_defs.c responsible for
emitting LFPC and STFPC instructions.  And since this frees up the FPC
save slot on the stack, adjust the stack layout accordingly.