Use Ints for fds in PRE and POST sys_close_range
The double_close_range test failed on riscv64-linux because the
close_range wrapper is using unsigned int and the ARG regwords
directly. Which causes the ARG2 == ~0U check to fail. Explicitly
using Int for the fd arguments fixes this. I am not clear on why
this was only an issue for the riscv port. It seems this patch is
OK for other arches (tested on amd64 and i386).