Bug 502324 - s390x: Fix memcheck false positives with TM/TMY
If the condition code of TM/TMY is generated in a different block than it
is used, memcheck can yield false positives for a partially initialized
value even if the checked bits are all defined.

Fix this by storing the operand ANDed with the mask in the flags thunk,
instead of the unmodified operand.  This enables memcheck to track the
defined bits correctly.