mount syscall param filesystemtype may be NULL
On Linux the mount syscall, depending on flags provided, the source,
type and data my be ignored.  We already don't check data and allow
source to be NULL.  Normally when type is ignored an application will
provide an empty string "".  But sometimes NULL is passed (like for
source).  So we now also allow type to be NULL to prevent false
positives.

Adjust the linux/scalar.c tests so the type param is still
unaddressable.

https://bugs.kde.org/show_bug.cgi?id=503914