Check dup2 oldfd before allowing the syscall
The dup201 LTP test fails with TFAIL: dup2(1024, 5) succeeded

That is because 1024 here is the soft file limit (so one higher than
the max number of fds). Valgrind raises the soft limit a little
internally to have a few private fds for itself. So this dup2 call
succeeds (and possibly dups and internal valgrind fd into the
newfd). We should check the oldfd before allowing the dup2 syscall,
like we already check the newfd.