nanomips specific changes for BZ 507033
Rework code to use Iop_ClzNat32 instead of the deprecated Iop_Clz32.

Iop_Clz32 is used to implement the CLZ insn which behaves naturally
when the input is 0:  CLZ(0) == 32
So it seems as if using Iop_Clz32 is wrong because it has undefined
behaviour when the input value is 0. However, the VEX pipeline
does this:

CLZ insn --ir--> Iop_Clz32 --isel--> NMun_CLZ --emit--> CLZ

So it all works out. Essentially the semantics of Iop_Clz32 were
redefined to be like Iop_ClzNat32.

In IR generation we can drop the special handling for CLO(0) because
Iop_ClzNat32 takes care of that.

Part of fixing https://bugs.kde.org/show_bug.cgi?id=507033