Allow for patching LTP sources
Sometimes there's an upstream LTP patch that helps testing
valgrind, but it's not yet part of the official LTP tarball.
In such cases it's helpful to be able to patch the LTP sources.
Attached patch allows for that.  It comes with a real life
example patch: LTP commit b62b831cf.