iropt-test: Constant folding for Iop_DivModU64to32 and Iop_DivModS64to32
Function get_selected_values has been tweaked to return more
"interesting" values for test generation. Namely those that,
when interpreted as a signed integer, are boundary (or near
boundary) values.

Also a bit of reorg because for the DivMod.. IROps we need to be a bit
more careful when generating testcases that do not trap or where the
result does not fit in 32 bit.

Part of fixing https://bugs.kde.org/show_bug.cgi?id=506211