Wrap the setdomainname syscall
The setdomainname() call sets the domain name to the value given in the
character array name.  The len argument specifies the number of bytes in
name.  (Thus, name does not require a terminating null byte.)

Declare a setdomainname wrapper in priv_syswrap-linux.h and hook it for
{amd64,arm,arm64,mips64,ppc32,ppc64,riscv64,s390x,x86}-linux using
LINX_ with PRE handler in syswrap-linux.c

https://bugs.kde.org/show_bug.cgi?id=369030