s390: Add bfp-emit.pl  (BZ 509572)
For a BFP insn X in the guest code the same insn will be emitted in
the jitted code. This does not hold universally but for most BFP
insns it does.
bfp-emit.pl contains a complete list of all BFP insns as of Principles
of Operations SA22-7832-14. It tests all insns for which the above
observation is true and ensures the emitted insn matches the insn in
the guest code.

Part of fixing https://bugs.kde.org/show_bug.cgi?id=509572