s390: Assorted BFP testsuite tweaks
Rename rounding-1 to bfp-306054 because that was the corresponding BZ.
Rename bfp-3 to bfp-muldiv as it is more expressive.
Only compile testcases for 128-bit BFP if the compiler supports -mlong-double-128.
Add -mlong-double-128 to the compile flags for the files that need it.

This concludes fixing https://bugs.kde.org/show_bug.cgi?id=509572