s390x: Fix inline assembly for STFLE
The inline assembly for emitting the STFLE instruction changes GPR 0, but
fails to mention it in the clobber list.  Depending on the compiler's
mood, this may result in wrong code.

A simple fix would be to just add "0" to the clobber list, but for
readability it also helps to move the inline assembly to a separate
function.  So do that.  In that function, use an "asm" variable for GPR 0
instead of moving the value back and forth between registers.