FreeBSD regtest: add a prereq for bug452274
This test tends to hang when running in a VM. The problem is that
the test spams both syscalls (write) and signals (SIGVTALRM from
setitimer). The test halts after 100 signals have been handled.
Under VirtualBox the timer signals get starved by the syscalls
and the limit of 100 doesn't get reached (at least not in any
reasonable amount of time).